+++ packages/composer/src/components/Preview/Proofreader.tsx
@@ -3,8 +3,9 @@ import debounce from 'lodash/debounce';
 import Dropdown from '@airbnb/lunar/lib/components/Dropdown';
 import ErrorMenu from '@airbnb/lunar/lib/components/Proofreader/ErrorMenu';
 import ControlBar from '@airbnb/lunar/lib/components/Proofreader/ControlBar';
+import { checkForAirbnbErrors } from '@airbnb/lunar/lib/components/Proofreader/helpers';
+import { NO_LOCALE } from '@airbnb/lunar/lib/components/Proofreader/constants';
 import { DEFAULT_LOCALE } from '@airbnb/lunar/lib/constants';
-import { checkForAirbnbErrors, NO_LOCALE } from '../../helpers/preview';
 import ComposerContext from '../../contexts/ComposerContext';
 import Renderer, { RendererProps } from './Renderer';
 import Window from './Window';
